Detail 컴포넌트 분리(#11)
