+++ src/components/detail/KakaoMap.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import * as S from './StyledDetail';
+
+function KakaoMap() {
+  return <S.detailKakaoMap id="map"></S.detailKakaoMap>;
+}
+
+export default KakaoMap;
