+++ src/components/detail/List.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import * as S from './StyledDetail';
+import { onClickSpotCreateMarker } from './KakaoMap';
+function List() {
+  return <S.detailPlaceList>List</S.detailPlaceList>;
+}
+
+export default List;
